refactor(api): move progression model
